bootstrap example/requirements.txt

bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
cryptography==3.4.7
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
mysqlclient==2.0.3
paramiko==2.7.2
protobuf==3.14.0
pycparser==2.20
PyNaCl==1.4.0
six==1.15.0
sshtunnel==0.4.0
Werkzeug==1.0.1
